add score support for Sage parameters adjust export report

+++ quicksearchprot/src/main/java/no/uib/probe/quicksearchprot/model/ResultScoreModel.java
@@ -76,7 +76,7 @@ public class ResultScoreModel {
 //            else {
 //            conf = "";
 //        }
-        return ("\t"+percentage + " %\t" + conf + "");
+        return ("\t"+((int)percentage) + " %\t" + conf + "");
     }
 
     public boolean isDefaultParameterValue() {
